Anagram kata: Changes anagram word collections to Set
Was previously storing the anagram dictionary as a hash of
{String => Array}, in which the key was a String of unique letters
making up the anagrams in the Array. The problem with this is that
it was possible to have the same word added multiple times. Using a Set
instead of an Array solves this issue w/o requiring any changes to the
code.

The dictionary now uses a Hash of {String, Set} internally, though it
should be noted that the #find_for method still returns a sorted Array.
(Set#sort --> Array)