Anagrams: Fixes (somewhat) bug w/ accented chars
The previous version of the method to determine the :dict key for
a given word, `#key_for(word)`, had a bug in which stripped out
accented versions of characters. Thus, 'résumé' would be sorted under
'mrsu' instead of 'eemrsu'.

The right solution would probably be to use `I18n::transliterate` but
I need to get a bit more familiar with this lib. Was getting an error
trying to load the `:en` locale.

As a stopgap while I figure out I18n, I added a few `gsub` steps to the
pipeline that ought to catch 90 percent of the accented chars.