Triangle Factors: Implements '.first_with_factor_count'
Exercise: Find the first number in the triangle sequence with
given number (or greater) of factors.

Taken from Project Euler problem 12, in which one is supposed to
find the first triangle with 500 factors.

So... this solution ain't cutting it. The algorithm works great
for low numbers, but performance sucks looking for factor counts
beyond a 100. I suspect this is do to the easy, but rather
ineffeciet factor searching algorythm.

TODO: Replace the current factoring expression with a better method.