Triangle Factors: Improves speed performance
My original solution used a simple, brute-force algorithm
to find all the factors of each triangle number: iterate
from 1 to sqrt of t, check for % == 0. This worked when
T was relatively small, but began to choke when searching for
T numbers with more than a hundred factors.

I replaced the simple factoring expression with a method that
begins with the prime factors (via Prime.prime_division), then
builds up the non-prime factors from there.

The previous version failed to find a T number with 500 factors
after more than 2 hours.  The current version finds it in less
than 2 milliseconds.