Fib Digit Count:
One of the problems from Project Euler is to determine the index
of the first number in the Fibonacci sequence containing 1000 digits.

Implemented a method that will take a count, C, and quickly return
the index of the sequence that has at least C digits. When searching
for the first Fib number with 10,000 digits, it takes ~ 0.1s.