Gilded Rose: Adds initial code to practice the Gilded Rose Kata
Added a class & matching test for the Gilded Rose refactoring kata.

The code here is based on the version Sandi Metz used in her talk,
"All the Little Things". It is slightly different than Jim Weirich's
version (itself a variation of a C# example). Jim used an unbound method,
`update_quality(item)` that took a Struct with three attributes: a `name`,
a `quality` metric, and a `sell_in` value representing the number of
days remaining before the item should be sold. In Sandi's version, the
method and the attributes belong to instances of a `GildedRose` class
(before refactoring). In both versions the `update_quality` method
(which Sandi call `tick`) is fifty or so lines of tangled, nested if
statements, that beg to be refactored into something more manageable.

I've also implemented the test Jim included with his version, adjusted
to accommodate using the GildedRose class, and ported from somewhat
dated rspec-given code to minitest-spec.  All the tests for the
existing code is passing. The tests for the new "Conjured Item" code is
set to skip, so we can refactor "in the green" before adding it.

See the README for a detailed description of the exercise.