Gilded Rose: Refactors from "small methods" to "small classes"
Following Sandi's advice from "All the Little Things", I tried to
refactor the Gilded Rose class into multple small classes. It was
a pretty easy follow on from the "small methods" refactoring done
in the previous step. I went back and watched the video of her
talk again and my solution ended up pretty close to hers. Essentially,
the GildedRose class becomes module that is a factory of the various
Item subclasses. I did deviate from her solution in a couple places.
First, I overrode the .new method on the module. This did feel a bit
smelly to me, but otherwise it would have required changing the tests
and I interpreted that as cheating. (Sandi use a .for method instead.
Presumably this required changes to the tests, but she didn't show
that.) Second, she replaced the case statement that maps names to
classes with a config Hash. Her solution assumes that all normal items
are called "Normal". I think this is an oversight. The tests (fm where?
not Jim's, but probably based on them.) assigned "Normal" to the names
of the object, but that appears to be a generic. If you read the
original problem description (and look at the original code) the
implication is that "normal" refers to an object with any name that
doesn't match one of the special cases. I also wanted to allow for
different kinds of "Backstage pass to ...." and "Conjured ....". This
requires matching on patterns rather than whole strings. With all that
in mind, I think sticking with the case statement is best.

This refactoring took the overall flog score down from 45.8 to 45.0,
The method average down from 6.5 to 5.0. The most complex method remains
at 10.9