Euler: Solves Problem 29
Determines the number of distinct powers in a**b, for all permutations
where a and b represent the set of integers from 2 to 100.

Implemented two methods to calculate the solution: JMath.power_permutation(range)
and JMath.distinct_powers(range).

Very happy with this solution. It uses only two single-line methods
that are relatively clear and runs in ~ 0.02 seconds.