Euler: Solves problem 37 - Truncatable Primes
The problem asked for the sum of the all the prime numbers, which
when truncatated from left-to-right, and from right-to-left, produce
only prime numbers.

It was noted that there are exactly 11 numbers in the set and that it
does not include single digit primes (2, 3, 5, 7).

To reach a solution, I use a few simple methods:

JMath.trunks(n) => returns the set of truncated numbers from the given
number

JMath.truncatable_primes => iterates through the set of prime numbers
until if finds 11 without any non-prime trunks.

Not sure if this is the most efficient solution, but it's fairly clear
and took about 2 seconds to complete. Not great, but acceptable.