Euler: Solves Problem 39 - Integer Right Triangles
The problems asks us to find the perimeter length less than or equal
to 1000 for which there are the greatest number of integer right
triangles.

Solved with two methods. First, JMath.right_triangles(perimeter), which returns
an Array of Arrays containing all the integer right triangle sets with
the given perimeter. Next, JMath.most_right_triangles(limit) returns the
number, up to the limit, with the greatest number of integer right
triangles.

The solution works, but I'm not happy with it. The right_triangles
method uses nested iteration to loop through the possible a, b, c
combinations for the given perimeter. This seems very inefficient and
will obviously degrade horribly as the size of p increases.

Will revisit to see if I can't produce a better algorithm.