Solves Project Euler Problems 18 and 67
Project Euler problem 18 asks presents set of numbers arranged in
a pyramid shape (one number in the first row, two in the second,
three in the third, etc.) and asks us to find the highest sum possible
following a path of adjacent numbers from top to bottom.

Problem 67 asks the same question, but given pyramid is 100 rows
high.

To solve these problems, I implemented a class, `Pyramid` with a
class method, `.maximum_path_sum(p)` that calculates the answer.

I also created a data class, `PyramidSets`, with three constants, `P1`,
`P2`, and `P3`. The first constant hold a small data set with a known
answer to use for testing. The other two hold the data sets for
Problem 18 and 67, respectively.

The tests for the Pyramid class (`test/pyramid_tests.rb`) include two
assertions that validate that the method works properly, and one
assertion that it operates quickly.

All tests green and answers verified on PE site.