Solves Project Euler Problem 19 - Counting Sundays
Problem 19 asks us to count the Sundays that occurred on the 1st of the
month during the 20th century.

Created `Sundays` class with method `.firsts_of_months_between(start,
finish)`. Using the Date class, it finds the next Sunday at or after
the given start dates, then checks each Sunday up to the finish date
to see if it's the 1st of a month.